ETWeb/requirements.txt
# Core Django packages
Django>=3.0.8,<4.0
djangorestframework>=3.11.0
django-environ>=0.4.5
django-rest-passwordreset>=1.1.0
django-heroku>=0.3.1

# Database
dj-database-url>=0.5.0
psycopg2-binary>=2.8.6

# Channels (WebSocket support) - without Redis
channels>=2.4.0,<4.0
daphne>=2.5.0

# ASGI and async support
asgiref>=3.2.0

# Image processing
Pillow>=9.0.0

# Utilities
pytz>=2021.1
whitenoise>=5.3.0
six>=1.16.0

# Supabase integration
supabase>=2.3.4
gotrue>=2.4.2
